pyproject.toml
[project]
name = "seliseblocks-genesis"
version = "0.1.0"
description = "Reusable FastAPI utilities like middlewares, singletons, and more."
authors = [{name="Selise Blocks", email="mostafizur.rahman@selisegroup.com"}]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-dotenv>=0.19.0",
    "azure-identity>=1.7.0",
    "azure-keyvault-secrets>=4.3.0",
    "redis>=4.0.0",
    "motor>=2.5.0",
    "structlog>=21.1.0",
    "pydantic>=2.11.5",
    "pymongo>=4.13.0",
    "opentelemetry-api>=1.33.1",
    "opentelemetry-sdk>=1.33.1",
    "opentelemetry-instrumentation-fastapi>=0.54b1",
    "opentelemetry-instrumentation-redis>=0.54b1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["blocks_genesis"]

